requirements.txt
asgiref==3.7.2
crispy-bootstrap5==2023.10
Django==5.0.1
django-crispy-forms==2.1
djangorestframework==3.14.0
et-xmlfile==1.1.0
openpyxl==3.1.2
pillow==10.2.0
pytz==2024.1
sqlparse==0.4.4
typing_extensions==4.9.0
